+++ assets/js/tecart-search.js
@@ -1,13 +1,13 @@
-// Search Bar & Toggle
-$('.tecart-search-toggle').on('click', function() {
-    // use classes and parent to be able to use search more than once on one page
-    $(this).parent().find('.tecart-search-form').toggle('display: inline-block');
-    $("input.tecart-search-box").focus();
-});
-// End - Search Bar & Toggle
-
 // Perform Search
 $(document).ready(function(){
+
+    // Search Bar & Toggle
+        $('.tecart-search-toggle').on('click', function() {
+            // use classes and parent to be able to use search more than once on one page
+            $(document).find('.tecart-search-form').toggle('display: inline-block');
+            $("input.tecart-search-box").focus();
+        });
+    // End - Search Bar & Toggle
 
     $.ajaxSetup({
         cache: false